feat: Add public API structs and templated function
Proposed Changes
================

This MR contains the Public API and the public structs required for the
Public API. It also contains a refactored build script with added
examples, unit tests, integration tests, and benchmark tests.

In addiction, there is also a makefile to simplify the commands to
compile, run the examples, etc. More commands can be added according to
the needs.