feat: Add CRC logic for Teltonika use case
Proposed Changes
================
Teltonika trackers when sending binary data via TCP transport protocol,
the CRC is in the payload. To verify when decoding a packet or even
generate a CRC when encoding one, this module was added.

The algorithm used in Teltonika is the CRC16/IBM, so only this algorithm
was supported.

Besides this addition, it also changes the tests location from `/tests/`
to `/src/tests/` to be able of testing the internal code.

Another addition was the usage example code in the documentation of the
module.