feat: Add AVL IO element parser (encode and decode)
Proposed Changes
================

Teltonika binary data format contains a AVL IO Element part that is
added in every AVL Data. This PR adds a parser for this data, containing
an encoder and a decoder. Supported codecs are Codec 8, 8E, and 16. It
is thread-safe.

In order to enable and facilitate the parser, it was also added a
ByteBuffer implementation without limitation of capacity allocating
memory on demand depending on the elements added to the buffer. This
implementation uses reflection to get bytes from the buffer but also to
put bytes into the buffer.

Changed the public struct AVL IO Element to contain Array Hash Maps to
maintain the order of insertion.

It was also updated the build config to support Zig version `0.15.1`.


**Dependencies Bumps**:

* zbench => from `0.10.1` to `0.11.2`.