feat: Add AVL Data parser
Proposed Changes
================

Teltonika binary data format contains AVL Data. This PR adds a parser
for this data, containing an encoder and a decoder. It is thread-safe.

* Added interfaces to `Crc`, `AvlIoElementParser`, and `AvlDataParser`.

* Restructured the build file to add the module dependencies to the
tests (unit, integration, and benchmark tests).

* Added `ByteBuferCombinedError` that combines ByteBuffer errors and
Allocator errors.

* Added an util to convert coordinate data from/to teltonika format
to/from floating coordinate format.

In terms of the tests structure, it was changed somes bits. It was added
a helpers folder, and also the naming of the files, when stored in tests
folder the names start with `test_*`, when files that are helpers starts
with `helper_*`.


**Dependencies Additions**:

* interface => `0.0.2` version.